Refactor credential_manager_ios package: remove cbor dependency and move AttestationParser to a new utils file
- Removed the `cbor` dependency from `pubspec.yaml` in both `credential_manager` and `credential_manager_ios` packages.
- Moved the `AttestationParser` class to a new `parser.dart` file within the `utils` directory for better organization and modularity.
- Updated imports accordingly to reflect the new structure.
