+++ packages/credential_manager/pubspec.yaml
@@ -20,7 +20,6 @@ dependencies:
   credential_manager_ios:
     path: ../credential_manager_ios
   plugin_platform_interface: ^2.0.2
-  cbor: ^6.3.5
 
 dev_dependencies:
   flutter_test:
