+++ packages/credential_manager_ios/lib/credential_manager_ios.dart
@@ -3,3 +3,4 @@ library credential_manager_ios;
 export 'src/credential_manager_ios.dart';
 export 'src/credential_manager_ios_plugin.dart';
 
+export 'src/utils/parser.dart';
