+++ packages/credential_manager_ios/lib/src/utils/parser.dart
