+++ packages/credential_manager_ios/pubspec.yaml
@@ -2,7 +2,6 @@ name: credential_manager_ios
 description: iOS implementation of the credential_manager plugin using Keychain and Autofill.
 version: 2.0.5
 homepage: https://github.com/Djsmk123/flutter_credential_manager_compose
-publish_to: none
 
 environment:
   sdk: '>=3.1.3 <4.0.0'
@@ -11,8 +10,7 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  credential_manager_platform_interface:
-    path: ../credential_manager_platform_interface
+  credential_manager_platform_interface: ^2.0.6
   plugin_platform_interface: ^2.0.2
   cbor: ^6.3.5
